[r2r] Lightning Swaps Locktimes + Lightning Maker Swap (#1557)
* use maker_lock_duration as min_final_cltv_expiry for maker lightning payment

* fix an issue where an invoice payment were not saved in db by the side that generated the invoice, add unit tests for this case

* use max_total_cltv_expiry_delta for send_taker_payment

* Make taker swap locktime when LightningCoin is Taker 24 hours and maker swap locktime for other coin 1.5x taker swap locktime

* validated min_final_cltv_expiry, add invoice expiry to swap invoices

* Some fixes and refactors for lightning swaps locktimes

* retry taker payment on multiple routes

* lightning maker swap

* refactor test_lightning_swaps

* fix rustfmt warnings

* remove unneeded todos and write better todos for next sprints

* fix wasm tests

* fix max_total_cltv_expiry_delta in send_taker_payment to use blocks

* Review fixes: use MmError::err, fix long todo msgs, avoid using error_log_with_msg when we need to log info too

* Review fixes: use casting for created_at/updated_at instead of TryInto

* Review fixes: fix some error messages,other refactors

+++ mm2src/mm2_main/src/wasm_tests.rs
@@ -22,8 +22,8 @@ fn wasm_start(ctx: MmArc) {
 /// and then immediately stops the nodes to check if `MmArc` is dropped in a short period.
 async fn test_mm2_stops_impl(
     pairs: &[(&'static str, &'static str)],
-    maker_price: i32,
-    taker_price: i32,
+    maker_price: f64,
+    taker_price: f64,
     volume: f64,
     stop_timeout_ms: u64,
 ) {
@@ -75,7 +75,7 @@ async fn test_mm2_stops_immediately() {
     register_wasm_log();
 
     let pairs: &[_] = &[("RICK", "MORTY")];
-    test_mm2_stops_impl(pairs, 1, 1, 0.0001, STOP_TIMEOUT_MS).await;
+    test_mm2_stops_impl(pairs, 1., 1., 0.0001, STOP_TIMEOUT_MS).await;
 }
 
 #[wasm_bindgen_test]
@@ -85,8 +85,8 @@ async fn trade_base_rel_electrum(
     bob_priv_key_policy: Mm2InitPrivKeyPolicy,
     alice_priv_key_policy: Mm2InitPrivKeyPolicy,
     pairs: &[(&'static str, &'static str)],
-    maker_price: i32,
-    taker_price: i32,
+    maker_price: f64,
+    taker_price: f64,
     volume: f64,
 ) {
     let coins = json!([rick_conf(), morty_conf(),]);
@@ -162,5 +162,5 @@ async fn trade_test_rick_and_morty() {
     let bob_policy = Mm2InitPrivKeyPolicy::Iguana;
     let alice_policy = Mm2InitPrivKeyPolicy::GlobalHDAccount(0);
     let pairs: &[_] = &[("RICK", "MORTY")];
-    trade_base_rel_electrum(bob_policy, alice_policy, pairs, 1, 1, 0.0001).await;
+    trade_base_rel_electrum(bob_policy, alice_policy, pairs, 1., 1., 0.0001).await;
 }
